- Merge generic and Core controls - Add breadscumbs in element doc pages
