+++ mkdocs.yml_template
@@ -26,8 +26,6 @@ nav:
           - "Visual Elements":
              - manuals/gui/viselements/index.md
              - "Controls": manuals/gui/viselements/controls.md
-             - "Core Back-end Controls":
-                - manuals/gui/corelements/index.md
              - "Blocks": manuals/gui/viselements/blocks.md
           - "Binding variables": manuals/gui/binding.md
           - "Callbacks": manuals/gui/callbacks.md
